build: fix 32-bit build by using uint64_t for to_u64_ceil max parameter
On 32-bit systems, size_t is only 32 bits, so passing UINT64_C(0x100000000)
(2^32) would overflow to 0. Change the parameter type to uint64_t.

+++ src/script/val64.cpp
@@ -240,7 +240,7 @@ void Val64::append_one()
     charvec_change_end();
 }
 
-uint64_t Val64::to_u64_ceil(size_t max, size_t &varcost) const
+uint64_t Val64::to_u64_ceil(uint64_t max, size_t &varcost) const
 {
     uint64_t v;
